IControlCalibration interface cleanup and directory refactoring. Fixes issue #1763 https://github.com/robotology/yarp/issues/1763 Bump version 3.0.101.
