+++ CMakeLists.txt
@@ -23,7 +23,7 @@ endif()
 
 # Main project
 project(YARP
-        VERSION 3.0.100
+        VERSION 3.0.101
         LANGUAGES C CXX)
 set(PROJECT_DESCRIPTION "YARP: A thin middleware for humanoid robots and more")
