+++ doc/release/v3_1_0.md
@@ -15,7 +15,11 @@ New Features
 * `Vocab` is now a namespace, and `yarp::os::createVocab` replaces `VOCAB1\2\3\4`
   that are now deprecated.
 
-### YARP_sig
+#### YARP_DEV
+
+* Class IControlCalibration refactored.
+  
+#### YARP_sig
 
 * `Vector` made typedef of `VectorOf<double>`(#1598).
