+++ src/libYARP_dev/include/yarp/dev/ControlBoardInterfacesImpl-inl.h
@@ -190,77 +190,6 @@ bool ImplementEncoders<DERIVED, IMPLEMENT>::getEncoderAccelerations(double *v)
     return ret;
 }
 
-////////////////////////
-// ControlCalibration Interface Implementation
-template <class DERIVED, class IMPLEMENT>
-ImplementControlCalibration<DERIVED, IMPLEMENT>::ImplementControlCalibration(DERIVED *y)
-{
-    iCalibrate= dynamic_cast<IControlCalibrationRaw *> (y);
-    helper = 0;
-    temp=0;
-}
-
-template <class DERIVED, class IMPLEMENT>
-ImplementControlCalibration<DERIVED, IMPLEMENT>::~ImplementControlCalibration()
-{
-    uninitialize();
-}
-
-template <class DERIVED, class IMPLEMENT>
-bool ImplementControlCalibration<DERIVED, IMPLEMENT>:: initialize (int size, const int *amap, const double *enc, const double *zos)
-{
-    if (helper!=0)
-        return false;
-
-    helper=(void *)(new ControlBoardHelper(size, amap, enc, zos));
-    yAssert (helper != 0);
-    temp=new double [size];
-    yAssert (temp != 0);
-    return true;
-}
-
-/**
-* Clean up internal data and memory.
-* @return true if uninitialization is executed, false otherwise.
-*/
-template <class DERIVED, class IMPLEMENT>
-bool ImplementControlCalibration<DERIVED, IMPLEMENT>::uninitialize ()
-{
-    if (helper!=0)
-    {
-        delete castToMapper(helper);
-        helper=0;
-    }
-
-    checkAndDestroy(temp);
-
-    return true;
-}
-
-template <class DERIVED, class IMPLEMENT>
-bool ImplementControlCalibration<DERIVED, IMPLEMENT>::calibrationDone(int j)
-{
-    int k=castToMapper(helper)->toHw(j);
-
-    return iCalibrate->calibrationDoneRaw(k);
-}
-
-template <class DERIVED, class IMPLEMENT>
-bool ImplementControlCalibration<DERIVED, IMPLEMENT>::calibrateAxisWithParams(int axis, unsigned int type, double p1, double p2, double p3)
-{
-    int k=castToMapper(helper)->toHw(axis);
-
-    return iCalibrate->calibrateAxisWithParamsRaw(k, type, p1, p2, p3);
-}
-
-template <class DERIVED, class IMPLEMENT>
-bool ImplementControlCalibration<DERIVED, IMPLEMENT>::setCalibrationParameters(int axis, const CalibrationParameters& params)
-{
-    int k = castToMapper(helper)->toHw(axis);
-
-    return iCalibrate->setCalibrationParametersRaw(k, params);
-}
-
 //////////////////////////////
 ///////////////// Implement
 template <class DERIVED, class IMPLEMENT>
