+++ src/libYARP_dev/include/yarp/dev/ControlBoardInterfacesImpl.h
@@ -23,6 +23,7 @@
 #include <yarp/dev/ImplementPWMControl.h>
 #include <yarp/dev/ImplementImpedanceControl.h>
 #include <yarp/dev/ImplementAxisInfo.h>
+#include <yarp/dev/ImplementControlCalibration.h>
 #include <yarp/dev/IPidControlImpl.h>
 #include <yarp/dev/IPositionControl2Impl.h>
 #include <yarp/dev/IVelocityControl2Impl.h>
